ltx-av: move vae_ltx.py to match Wan path structure
Move models/vae/ltx/vae_ltx.py → models/vae/vae_ltx.py to match
the Wan pattern (vae_wan2_1.py is at models/vae/ not in a subfolder).

Update all imports in pipeline, tests. Fix relative imports (4 dots → 3).
Remove empty ltx/ subfolder.
